# Latest/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.0.3",
    "title": "My User Project CRUD",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "127.0.0.1:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
    "paths": {
    "/line": {
      "get": {
        "tags": ["line"],
        "summary": "Request for line list",
		"parameters": [{
			"in":"query",
			"name":"lineNum",
			"schema":{"type":"string"},
			"required": false,
			"description": "Number of line."
		},{
			"in":"query",
			"name":"lineDir",
			"schema":{"type":"string"},
			"required": false,
			"description": "Direction of line."
		}],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Line"
            }
          }
        }
      }
    },
	"/stop": {
      "get": {
        "tags": ["stop"],
        "summary": "Request for stop list",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Stop"
            }
          }
        }
      }
    },
	"/linestop": {
      "get": {
        "tags": ["linestop"],
        "summary": "Request for line list",
		"parameters": [{
			"in":"query",
			"name":"lineNum",
			"schema":{"type":"string"},
			"required": true,
			"description": "Number of line."
		},{
			"in":"query",
			"name":"stopName",
			"schema":{"type":"string"},
			"required": true,
			"description": "Name of Stop."
		},{
			"in":"query",
			"name":"lineDir",
			"schema":{"type":"string"},
			"required": false,
			"description": "Direction of line."
		}],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LineStopResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Line": {
      "required": ["lineNum", "lineName", "stops", "lineGeoLoc"],
      "properties": {
        "lineNum": {
          "type": "string",
          "uniqueItems": true
        },
        "lineName": {
          "type": "string",
		  "uniqueItems": true
        },
        "lineDir": {
          "type": "string"
        },
        "stops": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "stopID": {
                "type": "integer"
              },
              "stopName": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "Stop": {
      "required": ["stopID", "stopName", "lines"],
      "properties": {
        "stopID": {
          "type": "integer",
		  "uniqueItems": true
        },
        "stopName": {
          "type": "string"
        },
        "lines": {
          "type": "array",
          "items": {
            "type": "SingleLine"
          }
        }
      }
    },
	"LineStopRequest": {
      "required": ["lineNum", "stopID", "stopName"],
      "properties": {
        "lineNum": {
          "type": "string",
		  "uniqueItems": true
        },
        "lineDir": {
          "type": "string"
        },
        "stopID": {
          "type": "integer"
        },
		"stopName": {
          "type": "string"
        }
      }
    },
	"LineStopResponse": {
      "required": ["lineNum", "ETA", "nextBusTime", "busGeoLoc"],
      "properties": {
        "lineNum": {
          "type": "string",
		  "uniqueItems": true
        },
        "lineDir": {
          "type": "string"
        },
        "ETA": {
          "type": "integer"
        },
		"nextBusTime": {
          "type": "string"
        },
		"busGeoLoc": {
          "type": "string"
        }
      }
    },
    "SingleStop": {
      "required": ["stopID", "stopName"],
      "properties": {
        "stopID": {
          "type": "string",
          "uniqueItems": true
        },
        "stopName": {
          "type": "string",
		  "uniqueItems": true
        }
      }
    },
	"SingleLine": {
      "required": ["lineNum", "lineName"],
      "properties": {
        "lineNum": {
          "type": "string",
          "uniqueItems": true
        },
        "lineName": {
          "type": "string",
		  "uniqueItems": true
        },
		"lineDir": {
          "type": "string",
		  "uniqueItems": true
        }
      }
    },
	"LinesList": {
      "required": ["lineList"],
      "properties": {
        "lineList": {
          "type": "array",
          "items": {
            "type": "SingleLine"
          }
        }
      }
    },
	"StopsList": {
      "required": ["stopList"],
      "properties": {
        "stopList": {
          "type": "array",
          "items": {
            "type": "SingleStop"
          }
        }
      }
    }
  }
}